# tickets-service/openapi.yaml
openapi: 3.0.3
info:
  title: Tickets Service API
  description: Support ticket management with escalation and AI features
  version: 1.0.0
servers:
  - url: http://localhost:8085
    description: Local development
...
paths:
...
  /api/v1/tickets/{id}/attachments:
    get:
      tags: [Attachments]
      summary: Get attachments
      operationId: getAttachments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Attachments list
    post:
      tags: [Attachments]
      summary: Add attachment
      operationId: addAttachment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Attachment added
...
  /api/v1/tickets/attachments/upload:
    post:
      tags: [Attachments]
      summary: Upload attachment
      operationId: uploadAttachment
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Attachment uploaded
  /api/v1/tickets/{id}/attachments/presigned-url:
    post:
      tags: [Attachments]
      summary: Generate presigned URL
      operationId: generatePresignedUrl
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Presigned URL
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT